requirements.txt
binance
pandas
pandas_ta
python-dotenv
requests 